rename NumericInputStatefulWrapper resolves #102
